Added WordFinder/Updated Prompter
Made the WordFinder class responsible for generating random words for passwords, be it through the user's word bank or randomly searching word files. Added the start of the Prompter functionality (not fully there yet). So far it only takes in one optional argument: "quick X" bypasses everything and generates X number of passwords.